test(guard): cover changed approval expiry path
